modify (pre)clustering workflows to process several ROF per TF

+++ Detectors/MUON/MCH/Base/include/MCHBase/PreCluster.h
@@ -29,11 +29,11 @@ namespace mch
 
 /// precluster minimal structure
 struct PreCluster {
-  uint16_t firstDigit; ///< index of first associated digit in the ordered vector of digits
-  uint16_t nDigits;    ///< number of digits attached to this precluster
+  uint32_t firstDigit; ///< index of first associated digit in the ordered vector of digits
+  uint32_t nDigits;    ///< number of digits attached to this precluster
 
   /// return the index of last associated digit in the ordered vector of digits
-  uint16_t lastDigit() const { return firstDigit + nDigits - 1; }
+  uint32_t lastDigit() const { return firstDigit + nDigits - 1; }
 
   void print(std::ostream& stream, gsl::span<const Digit> digits) const;
 };
